# docs/swagger.yaml
basePath: /
definitions:
  admin.Dashboard:
    properties:
      stock:
        type: integer
      user:
        type: integer
    type: object
  admin.LoginAdmin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  admin.User:
    properties:
      alamat:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
  admin.UserPoin:
    properties:
      alamat:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      poin:
        type: integer
    type: object
  dummy.AuthLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dummy.Bank:
    properties:
      amount:
        type: integer
      bankcode:
        type: string
      no_rekening:
        type: string
    type: object
  dummy.Customer:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      no_hp:
        type: string
      password:
        type: string
      pin:
        type: integer
      poin:
        type: integer
      token:
        type: string
    type: object
  dummy.DetailTransaction:
    properties:
      amount:
        type: integer
      an_rekening:
        type: string
      bank_provider:
        type: string
      createdat:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      jenis_transaction:
        type: string
      keterangan:
        type: string
      mitra_id:
        type: integer
      nomor:
        type: string
      poin_account:
        type: integer
      poin_redeem:
        type: integer
      status_poin:
        type: string
      status_transaction:
        type: string
      transaction_id:
        type: string
    type: object
  dummy.History:
    properties:
      id:
        type: integer
      keterangan:
        type: string
      status:
        type: string
      tanggal:
        type: string
      transaction_id:
        type: string
    type: object
  dummy.InputStockProduct:
    properties:
      stock:
        type: integer
    type: object
  dummy.Register:
    properties:
      email:
        type: string
      name:
        type: string
      no_hp:
        type: string
      password:
        type: string
    type: object
  dummy.StockProduct:
    properties:
      id:
        type: integer
      product:
        type: string
      stock:
        type: integer
    type: object
  mitra.AuthMitra:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  mitra.InputPoinMitra:
    properties:
      amount:
        type: integer
      idcustomer:
        type: integer
      idmitra:
        type: integer
      poin_account:
        type: integer
    type: object
  mitra.Mitra:
    properties:
      alamat:
        type: string
      email:
        type: string
      id:
        type: integer
      nama_toko:
        type: string
      password:
        type: string
    type: object
  mitra.MitraRegister:
    properties:
      alamat:
        type: string
      email:
        type: string
      nama_toko:
        type: string
      password:
        type: string
    type: object
  response.ResponseGetProduct:
    properties:
      code:
        type: integer
      message:
        type: string
      result:
        items:
          $ref: '#/definitions/dummy.StockProduct'
        type: array
    type: object
host: api-dummy.herokuapp.com
info:
  contact: {}
  description: Berikut API-Loyalty Point Agent
  title: API-Loyalty Point Agent
  version: "1.0"
paths:
  /v1/account:
    put:
      consumes:
      - application/json
      description: UpdateUser
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.Register'
      summary: UpdateUser
      tags:
      - User
  /v1/admin/dashboard:
    get:
      consumes:
      - application/json
      description: Dashboard Admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.Dashboard'
      summary: Dashboard Admin
      tags:
      - Admin
  /v1/admin/historycustomer:
    get:
      consumes:
      - application/json
      description: Customer history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.History'
      summary: Customer history
      tags:
      - Admin
  /v1/admin/login:
    post:
      consumes:
      - application/json
      description: Login Admin
      parameters:
      - description: Admin
        in: body
        name: LoginAdmin
        required: true
        schema:
          $ref: '#/definitions/admin.LoginAdmin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.LoginAdmin'
      summary: Login
      tags:
      - Admin
  /v1/admin/managecustomer:
    get:
      consumes:
      - application/json
      description: Manage Customer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.User'
      summary: Manage Customer
      tags:
      - Admin
  /v1/admin/managecustomer/{id}:
    put:
      consumes:
      - application/json
      description: Update customer data
      parameters:
      - description: id user
        in: path
        name: iduser
        required: true
        type: integer
      - description: User
        in: body
        name: Customer
        required: true
        schema:
          $ref: '#/definitions/admin.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.User'
      summary: Update customer data
      tags:
      - Admin
  /v1/admin/managecustomerpoint:
    get:
      consumes:
      - application/json
      description: Manage Customer Point
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.UserPoin'
      summary: Manage Customer Point
      tags:
      - Admin
  /v1/admin/managecustomerpoint/{id}:
    put:
      consumes:
      - application/json
      description: Update customer
      parameters:
      - description: id user
        in: path
        name: iduser
        required: true
        type: integer
      - description: User
        in: body
        name: Customer
        required: true
        schema:
          $ref: '#/definitions/admin.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.UserPoin'
      summary: Update customer point
      tags:
      - Admin
  /v1/detailhistory/{id}:
    get:
      consumes:
      - application/json
      description: History/transaction User
      parameters:
      - description: id detail history
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.DetailTransaction'
      summary: Detail History/transaction
      tags:
      - User
  /v1/history/{idcustomer}:
    get:
      consumes:
      - application/json
      description: History User
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id customer
        in: path
        name: idcustomer
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.History'
      summary: History
      tags:
      - User
  /v1/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dummy.AuthLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.Customer'
      summary: Login
      tags:
      - User
  /v1/mitra/inputpoin:
    post:
      consumes:
      - application/json
      description: Input Poin Mitra
      parameters:
      - description: InputPoinMitra
        in: body
        name: InputPoin
        required: true
        schema:
          $ref: '#/definitions/mitra.InputPoinMitra'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      summary: Input Poin Mitra
      tags:
      - Mitra
  /v1/mitra/login:
    post:
      consumes:
      - application/json
      description: Login Mitra
      parameters:
      - description: mitra
        in: body
        name: mitra
        required: true
        schema:
          $ref: '#/definitions/mitra.AuthMitra'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mitra.Mitra'
      summary: Login Mitra
      tags:
      - Mitra
  /v1/mitra/register:
    post:
      consumes:
      - application/json
      description: Register Mitra
      parameters:
      - description: MitraRegister
        in: body
        name: RegisterMitra
        required: true
        schema:
          $ref: '#/definitions/mitra.MitraRegister'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mitra.MitraRegister'
      summary: Register Mitra
      tags:
      - Mitra
  /v1/order/cashout:
    post:
      consumes:
      - application/json
      description: Order Cashout
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: inputdatacashout
        in: body
        name: InputDataCashout
        required: true
        schema:
          $ref: '#/definitions/dummy.Bank'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.Bank'
      summary: Order Cashout
      tags:
      - UserOrder
  /v1/order/emoney:
    post:
      consumes:
      - application/json
      description: Emoney user
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Order Emoney
      tags:
      - UserOrder
  /v1/order/paketdata:
    post:
      consumes:
      - application/json
      description: PaketData user
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Order PaketData
      tags:
      - UserOrder
  /v1/order/pulsa:
    post:
      consumes:
      - application/json
      description: Pulsa user
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Order Pulsa
      tags:
      - UserOrder
  /v1/register:
    post:
      consumes:
      - application/json
      description: Register user
      parameters:
      - description: Register
        in: body
        name: RegisterUser
        required: true
        schema:
          $ref: '#/definitions/dummy.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dummy.Register'
      summary: Register
      tags:
      - User
  /v1/stockproduct:
    get:
      consumes:
      - application/json
      description: Get Stock Product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseGetProduct'
      summary: Get Stock Product
      tags:
      - Product
  /v1/stockproduct/{id}:
    put:
      consumes:
      - application/json
      description: Update Stock Product
      parameters:
      - description: id product
        in: path
        name: id
        required: true
        type: integer
      - description: Update Stock Product
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/dummy.InputStockProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseGetProduct'
      summary: Update Stock Product
      tags:
      - Product
swagger: "2.0"